Revert "Revert "Update constants to suit new functions""
This reverts commit 0343ee73d8bba9ce1b2f13058d10761b501d806d.

+++ sashimi-webapp/src/database/data-modifier/dataDelete.js
@@ -9,17 +9,28 @@ const sqlCommands = new SqlCommands();
 export default class dataDelete {
   static constructor() {}
 
-  static deleteAllEntities() {
-    for (let index = 0; index < constants.ARRAY_ENTITIES_NAME.length; index += 1) {
-      sqlCommands.deleteTable(constants.ARRAY_ENTITIES_NAME[index]);
+  static deleteAllEntities(index) {
+    if (typeof Promise === 'function') {
+      return new Promise((resolve, reject) => {
+        if (index < constants.ARRAY_ENTITIES_NAME.length -1) {
+          sqlCommands.deleteTable(constants.ARRAY_ENTITIES_NAME[index])
+            .then(data => this.deleteAllEntities(index+1))
+            .catch(sqlError => reject(sqlError));
+          resolve(true);
+        }
+      });
+    } else {
+      throw new exceptions.PromiseFunctionNotDefined();
     }
   }
 
   static deleteFile(fileId) {
     if (typeof Promise === 'function') {
-      return new Promise((resolve, reject) => {
-        resolve(sqlCommands.deleteFile(fileId));
-      });
+      return new Promise((resolve, reject) =>
+        sqlCommands.deleteFile(fileId)
+          .then(data => resolve(data))
+          .catch(sqlError => reject(sqlError))
+      );
     } else {
       throw new exceptions.PromiseFunctionNotDefined();
     }
@@ -28,9 +39,11 @@ export default class dataDelete {
 
   static deleteFolder(folderId) {
     if (typeof Promise === 'function') {
-      return new Promise((resolve, reject) => {
-        resolve(sqlCommands.deleteFolder(folderId));
-      });
+      return new Promise((resolve, reject) =>
+        sqlCommands.deleteFolder(folderId)
+          .then(data => resolve(data))
+          .catch(sqlError => reject(sqlError))
+      );
     } else {
       throw new exceptions.PromiseFunctionNotDefined();
     }
